requirements.txt
flask-cors
flask
pandas
redis
tqdm
python-dotenv
gunicorn
pytz
requests